+++ examples/nextfs-starter/src/App/Template.fs
@@ -1,0 +1,12 @@
+module App.Template
+
+open Fable.Core
+open Feliz
+open NextFs
+
+[<ExportDefault>]
+let Template(props: TemplateProps) =
+    Html.div [
+        prop.className "min-h-screen"
+        prop.children [ props.children ]
+    ]